+++ src/main/java/com/example/webnovel/book/application/BookService.java
@@ -4,8 +4,8 @@ import com.example.webnovel.book.domain.book.Book;
 import com.example.webnovel.book.domain.book.Episode;
 import com.example.webnovel.book.domain.book.type.BookStatus;
 import com.example.webnovel.book.dto.BookResponse;
+import com.example.webnovel.book.dto.EpisodeResponse;
 import com.example.webnovel.book.infra.BookRepository;
-import com.example.webnovel.book.ui.EpisodeResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
