+++ src/test/java/com/example/webnovel/book/application/BookServiceTest.java
@@ -4,8 +4,8 @@ import com.example.webnovel.book.domain.book.Book;
 import com.example.webnovel.book.domain.book.Episode;
 import com.example.webnovel.book.domain.book.type.BookStatus;
 import com.example.webnovel.book.dto.BookResponse;
+import com.example.webnovel.book.dto.EpisodeResponse;
 import com.example.webnovel.book.infra.BookRepository;
-import com.example.webnovel.book.ui.EpisodeResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
